cpu.c: Add feature detection for AArch64 Neon FEAT_SHA3
This extension provides support for the BCAX ("bit-clear and xor") and
EOR3 ("exclusive-or 3-way") instructions.

This commit only adds detection logic for Apple Silicon and
getauxval-based platforms. At the time of writing Windows does not
provide a mechanism for detecting this feature. Android support is
covered by the getauxval version, since support for this feature was
only added after Android switched from android_getCpuFeatures to
getauxval.